+++ application/web/style/entry.css
@@ -1,0 +1,377 @@
+   /* __      _              
+  /__\ __ | |_ _ __ _   _ 
+ /_\| '_ \| __| '__| | | |
+//__| | | | |_| |  | |_| |
+\__/|_| |_|\__|_|   \__, |
+                    |___/  */
+
+
+.title-textarea {
+    /* background-color: #F3F3F3;
+    width: 20%;
+    min-height: 10px;
+    font-size: 20px;
+    padding: 10px;
+    margin-bottom: 10px;
+    box-sizing: border-box;
+    border: 3px solid #e2dede;
+    border-radius: 10px; */
+    background-color: var(--journal-entry);
+    padding: 10px;
+}
+
+.entry-textarea {
+    background-color: #e2dede;
+    width: 100%;
+    min-height: 400px;
+    font-size: 15px;
+    padding: 10px;
+    box-sizing: border-box;
+    border: none;
+    border-radius: 10px;
+}
+
+.new_entry {
+    color: #FFFFFF;
+    align-content: center;
+    font-family: var(--font-family-journals);
+    font-size: 20px;
+    font-weight: bold;
+}
+
+.entry-container {
+    display: flex;
+    flex-direction: row;
+    justify-content: center;
+    gap: 12px;
+}
+
+.cancel-save-container {
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+}
+
+#no-entry-text {
+    display: block;
+    margin-top: 10px;
+    font-size: 24px;
+}
+
+.home-entry-descriptor {
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+    padding-bottom: 10px;
+}
+
+.home-single-entry{
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    width: 100%;
+    padding: 10px 0px 2px; 
+    border-bottom: 2px solid var(--grey-text-color);
+}
+
+.home-single-entry-button {
+    font-size: 15px;
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+    width: 100%;
+    background-color: var(--journal-entry);
+    border: none;
+    cursor: pointer;
+
+}
+
+.home-list img[src="../web/assets/vdots-entry-black.png"]{
+    background-color: var(--journal-entry);
+    border: none;
+    margin-left: 20px;
+    width: 22px;
+    height: 22px;
+    cursor: pointer;
+}
+                    
+.title-textarea {
+    /* background-color: #F3F3F3;
+    width: 20%;
+    min-height: 10px;
+    font-size: 20px;
+    padding: 10px;
+    margin-bottom: 10px;
+    box-sizing: border-box;
+    border: 3px solid #e2dede;
+    border-radius: 10px; */
+    background-color: var(--journal-entry);
+    padding: 10px;
+}
+
+.entry-textarea {
+    background-color: #e2dede;
+    width: 100%;
+    min-height: 400px;
+    font-size: 15px;
+    padding: 10px;
+    box-sizing: border-box;
+    border: none;
+    border-radius: 10px;
+}
+
+.new_entry {
+    color: #FFFFFF;
+    align-content: center;
+    font-family: var(--font-family-journals);
+    font-size: 20px;
+    font-weight: bold;
+}
+
+.entry-container {
+    display: flex;
+    flex-direction: row;
+    justify-content: center;
+    gap: 12px;
+}
+
+.cancel-save-container {
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+}
+
+.no-entry-text{
+    font-family: var(--font-family);
+    margin-top: 10px;
+    font-size: 24px;
+}
+
+.no-entry-text{
+    font-family: var(--font-family);
+    margin-top: 10px;
+    font-size: 24px;
+}
+                    
+.no-display {
+    display: none;
+}
+
+.title-textarea {
+    /* background-color: #F3F3F3;
+    width: 20%;
+    min-height: 10px;
+    font-size: 20px;
+    padding: 10px;
+    margin-bottom: 10px;
+    box-sizing: border-box;
+    border: 3px solid #e2dede;
+    border-radius: 10px; */
+    background-color: var(--journal-entry);
+    padding: 10px;
+}
+
+.entry-textarea {
+    background-color: #e2dede;
+    width: 100%;
+    min-height: 400px;
+    font-size: 15px;
+    padding: 10px;
+    box-sizing: border-box;
+    border: none;
+    border-radius: 10px;
+}
+
+.new_entry {
+    color: #FFFFFF;
+    align-content: center;
+    font-family: var(--font-family-journals);
+    font-size: 20px;
+    font-weight: bold;
+}
+
+.entry-container {
+    display: flex;
+    flex-direction: row;
+    justify-content: center;
+    gap: 12px;
+}
+
+.cancel-save-container {
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+}
+
+
+/* The actual popup */
+.popup {
+    position: fixed;
+    z-index: 1;
+    left: 0;
+    top: 0;
+    width: 100%;
+    height: 100%;
+    overflow: auto;
+    background-color: rgba(0,0,0,0.4);
+    display: none;
+}
+.popup-content {
+    display: block;
+    background-color: var(--sliding-shelf);
+    margin: 15% auto;
+    padding: 20px;
+    min-width: 200px;
+    width: 13%;
+    height: 12%;
+    border-radius: 8px;
+    box-shadow: 2px 6px 5px 1px rgba(64,64,64,0.5);
+
+}
+.popup-content button {
+    display: block;
+}
+.show {
+    display: block;
+}
+
+.secondary {
+    background-color: var(--journal-entry);
+    color: var(--grey-text-color);
+    border-radius: 8px;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    cursor: pointer;
+    font-size: 15px;
+    min-height: 40px;
+    border: 1px solid;
+    font-family: var(--font-family-journals);
+}
+
+.primary {
+    background-color: var(--dark-green);
+    color: white;
+    border: none;
+    border-radius: 8px;
+    box-sizing: border-box;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    cursor: pointer;
+    font-size: 15px;
+    min-height: 40px;
+    font-family: var(--font-family-journals);
+}
+
+.title-entry {
+    margin: 0;
+    flex-grow: 1;
+}
+
+/* The div holding the buttons */
+.buttons {
+    display: flex;
+    gap: 10px;
+}
+
+#title-input {
+    font-size: 16px;
+    color: var(--dark-green);
+    background-color: var(--journal-entry);
+    border: none;
+    border-bottom: 2px solid var(--grey-text-color);
+    padding-left: 10px;
+    width: 200px;
+    outline: none;
+    transition: border-color 0.3s;
+}
+
+#title-input.placeholder {
+    color: var(--main-stroke-color);
+    border-bottom-color: var(--main-stroke-color);
+}
+
+#title-input.filled {
+    color: black;
+    border-bottom-color: var(--dark-blue);
+}
+
+#title-input:focus {
+    border-bottom-color: var(--dark-blue);
+}
+
+
+.discard-cancel-container {
+    display: flex;
+    flex-direction: row;
+    padding-top: 60px;
+    padding-left: 118px;
+    gap:10px;
+}
+
+/* Three dots */
+.three-dots-entry {
+    background-color: var(--journal-entry);
+    border: none;
+}
+
+.three-dots {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    cursor: pointer;
+    border: none; /* Ensure it looks like an icon, not a standard button */
+    background: none; /* Remove default button background */
+    padding: 0; /* Remove default button padding */
+}
+
+/* Default icon */
+.three-dots .dots-icon {
+    content: url('./assets/vdots-grey.svg');
+}
+
+/* Hover state */
+.three-dots:hover .dots-icon {
+    content: url('./assets/vdots-grey-active.svg');
+}
+
+.dropdown-menu {
+    display: block;
+    position: absolute;
+    background-color: var(--journal-entry);
+    border: 1px solid var(--icon-color);
+    border-radius: 8px;
+    box-shadow: 0 2px 5px rgba(0, 0, 0, 0.15);
+    margin-top: 10px;
+    z-index: 1000;
+}
+
+.dropdown-item {
+    padding: 10px 15px;
+    display: flex;
+    align-items: center;
+    cursor: pointer;
+}
+
+.dropdown-item:hover {
+    background-color: var(--icon-color);
+    border-radius: 5px;
+}
+
+.dropdown-icon {
+    margin-right: 10px;
+}
+
+.checkmark {
+    margin-left: auto;
+    visibility: hidden;
+}
+
+.dropdown-item.delete {
+    color: red;
+}
+
+.dropdown-item.delete:hover {
+    background-color: #f8d7da;
+}
