+++ application/web/style/global-variables.css
@@ -1,0 +1,17 @@
+   /* ___ _       _           _                   _       _     _           
+  / _ \ | ___ | |__   __ _| | /\   /\__ _ _ __(_) __ _| |__ | | ___  ___ 
+ / /_\/ |/ _ \| '_ \ / _` | | \ \ / / _` | '__| |/ _` | '_ \| |/ _ \/ __|
+/ /_\\| | (_) | |_) | (_| | |  \ V / (_| | |  | | (_| | |_) | |  __/\__ \
+\____/|_|\___/|_.__/ \__,_|_|   \_/ \__,_|_|  |_|\__,_|_.__/|_|\___||___/
+                                                                          */
+:root {
+    --journal-entry: #F3F3F3;
+    --sliding-shelf: #EBEBEB;
+    --main-stroke-color: #A1A1A1;
+    --grey-text-color: #404040;
+    --icon-color: #D7D7D7;
+    --font-family: 'Dosis', sans-serif;
+    --font-family-journals: 'Inter';
+    --dark-green: #58A399;
+    --dark-blue: #496989
+}
