+++ application/web/style/sidebar.css
@@ -1,0 +1,150 @@
+ /* __ _     _      _                
+/ _(_) __| | ___| |__   __ _ _ __ 
+\ \| |/ _` |/ _ \ '_ \ / _` | '__|
+_\ \ | (_| |  __/ |_) | (_| | |   
+\__/_|\__,_|\___|_.__/ \__,_|_|   
+                                   */
+sidebar {
+    width: 213px;
+    background-color: var(--sliding-shelf);
+    border-right: 1px solid var(--main-stroke-color);
+    padding: 10px;
+    box-sizing: border-box;
+    box-shadow: 2px 0 2px -2px rgba(0, 0, 0, 0.1);
+    display: flex;
+    flex-direction: column;
+    align-items: flex-start;
+    transition: width 0.3s ease;
+}
+
+.sidebar-header {
+    display: flex;
+    justify-content: flex-start;
+    width: 100%;
+}
+
+.sidebar-module {
+    display: flex;
+    flex-direction: column;
+    justify-content: left;
+    width: 100%;
+    transition: width 0.3s ease;
+    /* overflow: scroll;  */
+    /* to make journal list for too many journals */
+    overflow-y: scroll;
+}
+
+.sidebar-module::-webkit-scrollbar {
+    width: 0px; 
+}
+
+
+/* .sidebar-module .fa.fa-ellipsis-v{
+    display:none;
+    transition: width 0.3s ease;
+} */
+/* To counteract the swap from .fa to vdots, both will be here */
+.sidebar-module #sidebar-dots{
+    display:none;
+    transition: width 0.3s ease;
+    height: 20px;
+    width: 20px;
+}
+
+.sidebar-module .journal-entry #sidebar-dots:hover {
+    display: block;
+    content: url('./assets/vdots-journal-white-active.svg'); /* New image for hover */
+}
+
+button.new-journal {
+    background-color: var(--dark-green);
+    color: white;
+    border: none;
+    border-radius: 5px;
+    margin-bottom: 10px;
+    padding: 10px;
+    width: 100%;
+    box-sizing: border-box;
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    cursor: pointer;
+    font-size: 16px;
+    min-height: 40px;
+}
+
+button.collapse-button {
+    background-color: transparent;
+    border: none;
+    padding: 0;
+    border-radius: 5px;
+    margin-bottom: 10px;
+    cursor: pointer;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    height: 32px;
+    width: 32px;
+}
+
+button.new-journal i.icon {
+    margin-left: 10px;
+}
+
+button.new-journal img.journal-icon,
+img.collapse-button {
+    height: 20px;
+    width: 20px;
+}
+
+.journal-entry {
+    background-color: #cbcfce;
+    color: var(--grey-text-color);
+    border-radius: 5px;
+    padding: 10px;
+    width: 100%;
+    box-sizing: border-box;
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    cursor: pointer;
+    font-size: 16px;
+    min-height: 40px;
+    margin-bottom: 10px;
+    background: none;
+}
+
+
+.journal-entry i.icon {
+    margin-left: 10px;
+}
+
+.sidebar-collapsed {
+    width: 50px;
+}
+
+.toggled-sidebar-module {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    width: 100%;
+}
+
+.toggled-new-journal,
+.toggled-journal-entry {
+    transition: width 0.3s ease;
+    visibility: hidden;
+}
+
+.sidebar-collapsed button.collapse-button {
+    transform: rotate(180deg);
+}
+.edit-journal {
+    padding: 0;
+    border: none;
+    background: transparent;
+}
+.edit-journal > img {
+    cursor: pointer;
+}
